+++ deepinv/physics/mri.py
@@ -386,7 +386,7 @@ class MultiCoilMRI(MRIMixin, LinearPhysics):
         except ImportError:  # pragma: no cover
             raise ImportError(
                 "sigpy is required to simulate coil maps. Install it using pip install sigpy"
-            )  # pragma: no cover
+            )
 
         coil_maps = birdcage_maps(
             (n_coils,)
@@ -395,32 +395,54 @@ class MultiCoilMRI(MRIMixin, LinearPhysics):
         return torch.tensor(coil_maps).type(torch.complex64)
 
     @staticmethod
-    def estimate_coil_maps(y: Tensor, calib_size: int = 24) -> Tensor:
+    def estimate_coil_maps(
+        y: Tensor, calib_size: int = 24, use_cupy: bool = False
+    ) -> Tensor:
         """Estimate coil sensitivity maps using ESPIRiT.
 
         This was proposed in `ESPIRiT — An Eigenvalue Approach to Autocalibrating Parallel MRI: Where SENSE meets GRAPPA <https://onlinelibrary.wiley.com/doi/10.1002/mrm.24751>`_.
 
         Note this uses a suboptimal undifferentiable unbatched implementation provided by `sigpy`.
 
+        Optionally use `cupy` to accelerate on GPU, only if `cupy` is installed and a GPU is available.
+
         :param torch.Tensor y: multi-coil kspace measurements with shape [B,2,N,...,H,W] where N is coil dimension.
         :param int calib_size: optional square auto-calibration size in pixels, used by `sigpy`.
+        :param bool use_cupy: whether to attempt to use cupy for GPU acceleration.
         :return: torch.Tensor of coil maps of complex dtype and shape [B,N,...,H,W]
         """
         try:
             from sigpy.mri.app import EspiritCalib
-        except ImportError:
+            import sigpy as sp  # pragma: no cover
+        except ImportError:  # pragma: no cover
             raise ImportError(
                 "sigpy is required to estimate sens maps. Install it using pip install sigpy"
             )
 
-        return torch.from_numpy(
-            np.stack(
-                [
-                    EspiritCalib(yb, calib_size, show_pbar=False).run()
-                    for yb in MRIMixin.to_torch_complex(y).numpy()
-                ]
-            )
+        device = sp.Device(-1)
+        if use_cupy:  # pragma: no cover
+            try:
+                import cupy
+
+                device = sp.Device(0) if cupy.cuda.is_available() else sp.Device(-1)
+            except ImportError:
+                pass
+
+        maps = np.stack(
+            [
+                EspiritCalib(yb, calib_size, show_pbar=False, device=device).run()
+                for yb in MRIMixin.to_torch_complex(y).numpy()
+            ]
         )
+        if use_cupy:  # pragma: no cover
+            try:
+                import cupy
+
+                maps = cupy.asnumpy(maps)
+            except (ImportError, ModuleNotFoundError):
+                pass
+
+        return torch.from_numpy(maps)
 
 
 class DynamicMRI(MRI, TimeMixin):